Wow I actually worked on this thing again. Can you believe it?
Nodes can be renamed now. Double click the name to do it.
Added React name component for this
Added one more debug output button.

The engine is set up to handle the renaming but the UI is not...

After renaming, the knobs are all reset and no longer represent the values
assigned to the engine components. This is going to require an overhaul to the
UI to fix. I will likely end up fixing it by creating a new interface layer
between the UI and the engine that can be read from and written to by both.